Started zoom

+++ src/com/scs/physicsspacewar/entity/components/IDamagable.java
@@ -2,6 +2,6 @@ package com.scs.physicsspacewar.entity.components;
 
 public interface IDamagable {
 
-	void damage(float amt);
+	void damage(int amt);
 	
 }
